feat(build): update Makefile
Updated the Makefile to use a generic binary name in the build process. Instead of using "magistrala-$(1)" as the binary name, now it uses "$(1)". This change allows for easier customization and avoids hard-coding specific binary names for each service.
Revert Dockerfile to use a generic binary name in the build process. Instead of using "magistrala-$SVC" as the binary name, now it uses "$SVC". This change aligns with the updated Makefile and ensures consistency between the build process and the Docker image.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>