feat(workflow): check for generated files
The commit modifies the check-proto github workflow to check-generated-files. It checks for changes in specific paths and verifies the protobuf files and generated mocks. The code also includes checking if the SDK mock is up to date, installing a tool, generating mocks, and displaying an error message if the mocks are different.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>