fix(docker): Update container name and restart policy for RabbitMQ
This commit fixes the container name and restart policy for the RabbitMQ service in the docker-compose file. The container name is changed to "magistrala-rabbitmq" and the restart policy is set to "on-failure". This ensures that the RabbitMQ container has a unique and identifiable name and will automatically restart in case of failure.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>