test(evnts): Test `TestUnavailablePublish`
This commit adds a new test case to the publisher. It also removes a method that is no longer needed. The code in this file handles publishing and subscribing to events using RabbitMQ and Redis. It includes logic for pausing and unpausing a container, generating random events, and checking the number of received events. Furthermore, an argument is added to the NewPublisher function to allow specifying an interval for checking unpublished events.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>